version 2

+++ server/controllers/api.js
@@ -1,12 +1,14 @@
 const Post = require("../models/posts");
+const Comment = require("../models/comment");
 const fs = require("fs");
+
 
 module.exports = class API {
   // fetch all posts
   static async fetchAllPost(req, res) {
     console.log("fetchAllPost hits")
     try {
-      const posts = await Post.find();
+      const posts = await Post.find().populate('comments');
       res.status(200).json(posts);
     } catch (err) {
       res.status(404).json({ message: err.message });
@@ -20,7 +22,7 @@ module.exports = class API {
     console.log("fetchPostByID hits")
     const id = req.params.id;
     try {
-      const post = await Post.findById(id);
+      const post = await Post.findById(id).populate('comments');
       res.status(200).json(post);
     } catch (err) {
       res.status(404).json({ message: err.message });
@@ -32,7 +34,13 @@ module.exports = class API {
   // create a post
   static async createPost(req, res) {
     console.log("createPost hits")
-    const post = req.body;
+    const { title, category, content } = req.body;
+    let postedBy = req.user.username;
+
+    let post = {
+      title, category, content, postedBy
+    }
+
     try {
       await Post.create(post);
       res.status(201).json({ message: " Post created successfully ! " });
@@ -49,6 +57,11 @@ module.exports = class API {
 
     const newPost = req.body;
     try {
+      let post = await Post.findById(id);
+      if (post?.postedBy !== req?.user?.username) {
+        return res.status(401).json({ message: "not authorized" });
+      }
+
       await Post.findByIdAndUpdate(id, newPost);
       res.status(200).json({ message: 'Post updated successfully!' });
     } catch (err) {
@@ -61,8 +74,58 @@ module.exports = class API {
     console.log("deletePost hits")
     const id = req.params.id;
     try {
+      let post = await Post.findById(id);
+      if (post?.postedBy !== req?.user?.username) {
+        return res.status(401).json({ message: "not authorized" });
+      }
+
       await Post.findByIdAndDelete(id);
       res.status(200).json({ message: 'Post deleted successfully!' });
+    } catch (err) {
+      res.status(404).json({ message: err.message });
+    }
+  }
+
+
+  // add a comment
+  static async addComment(req, res) {
+    console.log("addComment hits")
+    const id = req.params.id;
+
+    const { comment } = req.body;
+    let username = req.user.username;
+    try {
+      const newComment = new Comment({
+        author: username,
+        comment: comment
+      });
+      newComment.save((err, result) => {
+        if (err) {
+          res.status(400).json({ message: err.message });
+        } else {
+          Post.findById(id, (posterr, post) => {
+            if (posterr) {
+              res.status(400).json({ message: posterr.message });
+            } else {
+              post.comments.push(result);
+              post.save((saverr, result) => {
+                if (saverr) {
+                  res.status(400).json({ message: saverr.message });
+                } else {
+                  Post.findById(id).populate('comments')
+                    .then((result) => {
+                      const comments = result?.comments;
+                      res.status(200).json({ comments });
+                    })
+                    .catch((err) => {
+                      res.status(400).json({ message: err.message });
+                    });
+                }
+              });
+            }
+          })
+        }
+      })
     } catch (err) {
       res.status(404).json({ message: err.message });
     }
